stock-data-backend/requirements.txt
psycopg[binary,pool]
cryptography
python-dotenv
yfinance
apscheduler
pandas
numpy
pytest
pytest-mock
akshare

# Core web and ML dependencies
flask
uvicorn
torch
scikit-learn
sqlalchemy
gunicorn
pydantic

# New dependencies for sentiment analysis
requests
transformers
akshare